+++ niceview/__init__.py
@@ -20,7 +20,7 @@ from niceview.dataadapter import (
     lenient_model_load,
 )
 from niceview.widgets import CheckboxGroup, field_value, render_field, to_widget_value
-from niceview.modelform import ModelForm
+from niceview.modelform import FormAction, ModelForm
 from niceview.modelgrid import ModelGrid, ModelGridInlineEdit
 from niceview.editwrapper import EditFormWrapper, EditGridWrapper
 from niceview.modellist import ModelList
@@ -35,7 +35,7 @@ __all__ = [
     # Single widgets without a model
     'render_field', 'field_value', 'to_widget_value',
     # UI components
-    'ModelForm', 'CheckboxGroup',
+    'ModelForm', 'FormAction', 'CheckboxGroup',
     'ModelGrid', 'ModelGridInlineEdit',
     'EditFormWrapper', 'EditGridWrapper',
     'ModelList', 'DrillDownWrapper',